packages/rag/requirements-lock.txt
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.3
    # via
    #   fsspec
    #   instructor
    #   langchain-community
aiosignal==1.4.0
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.12.0
    # via
    #   google-genai
    #   httpx
    #   openai
appdirs==1.4.4
    # via ragas
async-timeout==4.0.3
    # via
    #   aiohttp
    #   langchain-classic
attrs==25.4.0
    # via aiohttp
certifi==2022.12.7
    # via
    #   httpcore
    #   httpx
    #   requests
cfgv==3.5.0
    # via pre-commit
charset-normalizer==3.1.0
    # via requests
click==8.3.1
    # via
    #   typer
    #   typer-slim
dataclasses-json==0.6.7
    # via langchain-community
datasets==4.5.0
    # via ragas
dill==0.4.0
    # via
    #   datasets
    #   multiprocess
diskcache==5.6.3
    # via
    #   instructor
    #   ragas
distlib==0.3.6
    # via virtualenv
distro==1.9.0
    # via
    #   google-genai
    #   openai
docstring-parser==0.17.0
    # via instructor
exceptiongroup==1.3.1
    # via anyio
filelock==3.12.0
    # via
    #   datasets
    #   huggingface-hub
    #   virtualenv
filetype==1.2.0
    # via langchain-google-genai
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
fsspec[http]==2025.10.0
    # via
    #   datasets
    #   huggingface-hub
google-auth[requests]==2.47.0
    # via google-genai
google-genai==1.59.0
    # via langchain-google-genai
h11==0.16.0
    # via httpcore
hf-xet==1.2.0
    # via huggingface-hub
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   datasets
    #   google-genai
    #   huggingface-hub
    #   langgraph-sdk
    #   langsmith
    #   openai
httpx-sse==0.4.3
    # via
    #   langchain-community
    #   langchain-naver
huggingface-hub==1.3.2
    # via datasets
identify==2.6.16
    # via pre-commit
idna==3.4
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
instructor==1.14.4
    # via ragas
jinja2==3.1.6
    # via instructor
jiter==0.11.1
    # via
    #   instructor
    #   openai
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==1.2.3
    # via
    #   -r requirements.txt
    #   ragas
langchain-classic==1.0.1
    # via langchain-community
langchain-community==0.4.1
    # via ragas
langchain-core==1.2.6
    # via
    #   langchain
    #   langchain-classic
    #   langchain-community
    #   langchain-google-genai
    #   langchain-naver
    #   langchain-openai
    #   langchain-text-splitters
    #   langgraph
    #   langgraph-checkpoint
    #   langgraph-prebuilt
    #   ragas
langchain-google-genai==4.2.0
    # via -r requirements.txt
langchain-naver==0.1.1
    # via -r requirements.txt
langchain-openai==1.1.7
    # via
    #   langchain-naver
    #   ragas
langchain-text-splitters==1.1.0
    # via langchain-classic
langgraph==1.0.5
    # via langchain
langgraph-checkpoint==3.0.1
    # via
    #   langgraph
    #   langgraph-prebuilt
langgraph-prebuilt==1.0.5
    # via langgraph
langgraph-sdk==0.3.1
    # via langgraph
langsmith==0.6.0
    # via
    #   langchain-classic
    #   langchain-community
    #   langchain-core
markdown-it-py==4.0.0
    # via rich
markupsafe==2.1.2
    # via jinja2
marshmallow==3.26.2
    # via dataclasses-json
mdurl==0.1.2
    # via markdown-it-py
multidict==6.7.0
    # via
    #   aiohttp
    #   yarl
multiprocess==0.70.18
    # via datasets
mypy-extensions==1.1.0
    # via typing-inspect
nest-asyncio==1.6.0
    # via ragas
networkx==3.4.2
    # via ragas
nodeenv==1.10.0
    # via pre-commit
numpy==1.26.4
    # via
    #   datasets
    #   langchain-community
    #   pandas
    #   pgvector
    #   ragas
    #   scikit-network
    #   scipy
openai==2.15.0
    # via
    #   instructor
    #   langchain-openai
    #   ragas
orjson==3.11.5
    # via
    #   langgraph-sdk
    #   langsmith
ormsgpack==1.12.1
    # via langgraph-checkpoint
packaging==25.0
    # via
    #   datasets
    #   huggingface-hub
    #   langchain-core
    #   langsmith
    #   marshmallow
pandas==2.0.0
    # via datasets
pgvector==0.4.2
    # via -r requirements.txt
pillow==12.1.0
    # via ragas
platformdirs==3.5.1
    # via virtualenv
pre-commit==4.5.1
    # via instructor
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
psycopg2-binary==2.9.11
    # via -r requirements.txt
pyarrow==22.0.0
    # via datasets
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pydantic==2.12.5
    # via
    #   -r requirements.txt
    #   google-genai
    #   instructor
    #   langchain
    #   langchain-classic
    #   langchain-core
    #   langchain-google-genai
    #   langgraph
    #   langsmith
    #   openai
    #   pydantic-settings
    #   ragas
pydantic-core==2.41.5
    # via
    #   instructor
    #   pydantic
pydantic-settings==2.12.0
    # via langchain-community
pygments==2.19.2
    # via rich
python-dateutil==2.8.2
    # via pandas
python-dotenv==1.2.1
    # via
    #   -r requirements.txt
    #   pydantic-settings
pytz==2023.3
    # via pandas
pyyaml==6.0.3
    # via
    #   datasets
    #   huggingface-hub
    #   langchain-classic
    #   langchain-community
    #   langchain-core
    #   pre-commit
ragas==0.4.3
    # via -r requirements.txt
regex==2026.1.15
    # via tiktoken
requests==2.32.5
    # via
    #   datasets
    #   google-auth
    #   google-genai
    #   instructor
    #   langchain-classic
    #   langchain-community
    #   langsmith
    #   requests-toolbelt
    #   tiktoken
requests-toolbelt==1.0.0
    # via langsmith
rich==14.2.0
    # via
    #   instructor
    #   ragas
    #   typer
rsa==4.9
    # via google-auth
scikit-network==0.33.5
    # via ragas
scipy==1.10.1
    # via scikit-network
shellingham==1.5.4
    # via
    #   huggingface-hub
    #   typer
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   google-genai
    #   openai
sqlalchemy==2.0.45
    # via
    #   langchain-classic
    #   langchain-community
tenacity==9.1.2
    # via
    #   google-genai
    #   instructor
    #   langchain-community
    #   langchain-core
tiktoken==0.12.0
    # via
    #   langchain-openai
    #   ragas
tqdm==4.67.1
    # via
    #   datasets
    #   huggingface-hub
    #   openai
    #   ragas
ty==0.0.12
    # via instructor
typer==0.21.1
    # via
    #   instructor
    #   ragas
typer-slim==0.21.1
    # via huggingface-hub
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   anyio
    #   exceptiongroup
    #   google-genai
    #   huggingface-hub
    #   langchain-core
    #   multidict
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typer
    #   typer-slim
    #   typing-inspect
    #   typing-inspection
typing-inspect==0.9.0
    # via dataclasses-json
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
tzdata==2023.3
    # via pandas
urllib3==1.26.15
    # via requests
uuid-utils==0.12.0
    # via
    #   langchain-core
    #   langsmith
virtualenv==20.23.0
    # via pre-commit
websockets==15.0.1
    # via google-genai
xxhash==3.6.0
    # via
    #   datasets
    #   langgraph
yarl==1.22.0
    # via aiohttp
zstandard==0.25.0
    # via langsmith
